% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/domain_info.R
\name{certificate_info}
\alias{certificate_info}
\title{Get Certificate Information for a Domain}
\usage{
certificate_info(token, domain, ...)
}
\arguments{
\item{token}{a token generated with \code{\link{whoapi_token}}}

\item{domain}{a domain name}

\item{...}{further arguments to pass to httr's GET.}
}
\description{
\code{certificate_info} retrieves information
about a domain's SSL certificate (if present).
}
\examples{
token <- whoapi_token("demokey")
whoapi_cert <- certificate_info(token, "whoapi.com")
}
\seealso{
\code{\link{whois_info}} for information about the domain
in general.
}

