% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domain_metadata.R
\name{domain_metadata}
\alias{domain_metadata}
\title{Retrieve Domain Metadata}
\usage{
domain_metadata(token, domain, ...)
}
\arguments{
\item{token}{a token generated with \code{\link{whoapi_token}}}

\item{domain}{a domain name}

\item{...}{further arguments to pass to httr's GET.}
}
\description{
\code{domain_metadata} retrieves
information about the content on a domain,
specifically the title and metadata description
fields from its home page
}
\examples{
token <- whoapi_token("demokey")
\dontrun{
metadata <- domain_metadata(token, "whoapi.com")
}
}

