% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domain_metadata.R
\name{domain_search}
\alias{domain_search}
\title{Identify a Domain's Search Results Count}
\usage{
domain_search(token, domain, ...)
}
\arguments{
\item{token}{a token generated with \code{\link{whoapi_token}}}

\item{domain}{a domain name}

\item{...}{further arguments to pass to httr's GET.}
}
\description{
\code{domain_search} allows you to
quickly calculate, in an automated way, how many
search results Bing and Google return for a particular
domain.
}
\examples{
token <- whoapi_token("demokey")
\dontrun{
search_results <- domain_search(token, "whoapi.com")
}
}
\seealso{
\code{\link{domain_rank}} which looks more
generally at domain popularity according to Alexa and Google Pagerank
scores.
}

