% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/domain_info.R
\name{is_blacklisted}
\alias{is_blacklisted}
\title{Identify if a Domain is Blacklisted}
\usage{
is_blacklisted(token, domain, ...)
}
\arguments{
\item{token}{a token generated with \code{\link{whoapi_token}}}

\item{domain}{a domain name}

\item{...}{further arguments to pass to httr's GET.}
}
\value{
a list containing a boolean value, "blacklisted", which identifies whether a
domain is blacklisted by \emph{any} of the checked services, followed by a breakdown of what the
status is on each particular service.
}
\description{
\code{is_blacklisted} checks whether a domain is on prominent
spam blacklists (or not).
}

