% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{whoapi_token}
\alias{whoapi_token}
\title{Create a token for WhoAPI}
\usage{
whoapi_token(key, user_agent = NULL)
}
\arguments{
\item{key}{the API key to use.}

\item{user_agent}{the user agent to use. NULL by default, which will use the \code{whoapi} package user agent.}
}
\description{
This function allows the creation of connectors to whoapi, which can be passed
between function calls, providing consistent API keys and options for different calls.
}

