% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domain_info.R
\name{whois_info}
\alias{whois_info}
\title{Get Registration Information for a Domain}
\usage{
whois_info(token, domain, ...)
}
\arguments{
\item{token}{a token generated with \code{\link{whoapi_token}}}

\item{domain}{a domain name}

\item{...}{further arguments to pass to httr's GET.}
}
\description{
\code{whois_info} grabs detailed information about a domain's registration,
including (but not limited to) who it's registered to, what its status is, when it
was registered and when it expires.
}
\examples{
token <- whoapi_token("demokey")
\dontrun{
whoapi_whois_info <- whois_info(token, "whoapi.com")
}
}
\seealso{
\code{\link{is_taken}} for simpy determining if a domain is registered
at all, or \code{\link{certificate_info}} for information about a domain's
SSL certificates.
}

